@@ -1,5 +1,8 @@
+[project]
+requires-python = ">=3.12"
+
 [tool.black]
-target-version = ["py310"]
+target-version = ["py312"]
 extend-exclude = "/generated/"
 
 [tool.isort]
