@@ -1,10 +1,12 @@
 """The test for the snowtire binary sensor platform."""
+
 # pylint: disable=redefined-outer-name
 from typing import Final
 from unittest.mock import MagicMock
 
 import pytest
 from pytest import raises
+from pytest_homeassistant_custom_component.common import async_mock_service
 
 from custom_components.snowtire.binary_sensor import (
     SnowtireBinarySensor,
@@ -17,13 +19,18 @@ from custom_components.snowtire.const import (
     ICON_WINTER,
 )
 from homeassistant.components.weather import (
-    ATTR_FORECAST,
     ATTR_FORECAST_TEMP,
     ATTR_FORECAST_TEMP_LOW,
     ATTR_FORECAST_TIME,
     ATTR_WEATHER_TEMPERATURE,
+    SERVICE_GET_FORECASTS,
+    WeatherEntityFeature,
 )
-from homeassistant.const import CONF_PLATFORM, TEMP_CELSIUS, TEMP_FAHRENHEIT
+from homeassistant.const import (
+    ATTR_SUPPORTED_FEATURES,
+    CONF_PLATFORM,
+    UnitOfTemperature,
+)
 from homeassistant.core import HomeAssistant, State
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.util import dt as dt_util
@@ -88,9 +95,12 @@ async def test_async_added_to_hass(default_sensor):
 )
 async def test__temp2c(temp1, temp2):
     """Test temperature conversions."""
-    assert SnowtireBinarySensor._temp2c(temp1, TEMP_CELSIUS) == temp1
-    assert round(SnowtireBinarySensor._temp2c(temp1, TEMP_FAHRENHEIT), 2) == temp2
-    assert SnowtireBinarySensor._temp2c(None, TEMP_CELSIUS) is None
+    assert SnowtireBinarySensor._temp2c(temp1, UnitOfTemperature.CELSIUS) == temp1
+    assert (
+        round(SnowtireBinarySensor._temp2c(temp1, UnitOfTemperature.FAHRENHEIT), 2)
+        == temp2
+    )
+    assert SnowtireBinarySensor._temp2c(None, UnitOfTemperature.CELSIUS) is None
 
 
 async def test_async_update(hass: HomeAssistant, default_sensor):
@@ -109,30 +119,36 @@ async def test_async_update(hass: HomeAssistant, default_sensor):
     today_ts = int(today.timestamp() * 1000)
     day = days = 86400000
 
-    forecast = [
-        {
-            ATTR_FORECAST_TIME: today_ts - day,
-        },
-        {
-            ATTR_FORECAST_TIME: today,
-            ATTR_FORECAST_TEMP: 9,
-        },
-        {
-            ATTR_FORECAST_TIME: today_ts + day,
-            ATTR_FORECAST_TEMP_LOW: 1,
-            ATTR_FORECAST_TEMP: 8,
-        },
-        {
-            ATTR_FORECAST_TIME: today_ts + (MOCK_DAYS + 1) * days,
-        },
-    ]
+    forecast = {
+        MOCK_WEATHER_ENTITY: {
+            "forecast": [
+                {
+                    ATTR_FORECAST_TIME: today_ts - day,
+                },
+                {
+                    ATTR_FORECAST_TIME: today,
+                    ATTR_FORECAST_TEMP: 9,
+                },
+                {
+                    ATTR_FORECAST_TIME: today_ts + day,
+                    ATTR_FORECAST_TEMP_LOW: 1,
+                    ATTR_FORECAST_TEMP: 8,
+                },
+                {
+                    ATTR_FORECAST_TIME: today_ts + (MOCK_DAYS + 1) * days,
+                },
+            ]
+        }
+    }
+
+    async_mock_service(hass, CONF_WEATHER, SERVICE_GET_FORECASTS, response=forecast)
 
     hass.states.async_set(
         MOCK_WEATHER_ENTITY,
         "State",
         attributes={
             ATTR_WEATHER_TEMPERATURE: -1,
-            ATTR_FORECAST: forecast,
+            ATTR_SUPPORTED_FEATURES: WeatherEntityFeature.FORECAST_DAILY,
         },
     )
 
@@ -146,7 +162,7 @@ async def test_async_update(hass: HomeAssistant, default_sensor):
         "State",
         attributes={
             ATTR_WEATHER_TEMPERATURE: 9.9,
-            ATTR_FORECAST: forecast,
+            ATTR_SUPPORTED_FEATURES: WeatherEntityFeature.FORECAST_DAILY,
         },
     )
 
@@ -160,7 +176,7 @@ async def test_async_update(hass: HomeAssistant, default_sensor):
         "State",
         attributes={
             ATTR_WEATHER_TEMPERATURE: 10,
-            ATTR_FORECAST: forecast,
+            ATTR_SUPPORTED_FEATURES: WeatherEntityFeature.FORECAST_DAILY,
         },
     )
 
